synthetic app saving first data with all systems

+++ apps/synthetic-cpp/plot.sh
@@ -101,26 +101,47 @@ if [ "$PLOTID" == "2" ]; then
     files=
     NORMALIZE=1
     CORES=10
+    ZIPFS=0.85
+    # ZIPFS=0.5
 
     ## data
-    for runcfg in "noprio" "prio" "noprio+sc" "prio+sc" "prio+sc+bh" "prio+sc+bh+init"; do
+    # for runcfg in "noprio" "prio" "noprio+sc" "prio+sc" "prio+sc+bh" "prio+sc+bh+init"; do
+        # cfg="prio"
+    # for runcfg in "aifm" "fswap" "eden-sc" "eden+bh-sc" "eden" "eden+bh" "eden+bh-sc-prio"; do
+        # cfg="all"
+    # for runcfg in "aifm" "fswap"; do
+    #     cfg="aifm-fswap"
+    for runcfg in "aifm" "fswap" "eden-sc" "eden+bh-sc"; do
+        cfg="aifm-fswap-prio"
+    # for runcfg in "aifm" "fswap" "eden-sc" "eden+bh-sc" "eden" "eden+bh"; do
+    #     cfg="aifm-fswap-prio-sc"
+        
         LABEL=
         LS=
         CMI=
+        EDENDESC="goodruns"
+        # EDENDESC="zs0.5"
 
         case $runcfg in
-        "noprio")               pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=no; desc="rdma"; LS=dashed; CMI=0; LABEL="No-Prio";;
-        "prio")                 pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio";;
-        "noprio+sc")            pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=no; desc="rdma"; LS=dashed; CMI=0; LABEL="No-Prio(SC)";;
-        "prio+sc")              pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio(SC)";;
-        "prio+sc+bh")           pattern="04-07"; rmem=eden-bh; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio(SC+BH)";;
-        "prio+sc+bh+init")      pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma-noinit"; LS=solid; CMI=1; LABEL="Prio(SC+NoInit)";;
+        # "noprio")               pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=no; desc="rdma"; LS=dashed; CMI=0; LABEL="No-Prio";;
+        # "prio")                 pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio";;
+        # "noprio+sc")            pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=no; desc="rdma"; LS=dashed; CMI=0; LABEL="No-Prio(SC)";;
+        # "prio+sc")              pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio(SC)";;
+        # "prio+sc+bh")           pattern="04-07"; rmem=eden-bh; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma"; LS=solid; CMI=1; LABEL="Prio(SC+BH)";;
+        # "prio+sc+bh+init")      pattern="04-07"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc="rdma-noinit"; LS=solid; CMI=1; LABEL="Prio(SC+NoInit)";;
+        "aifm")                 rmem=aifm; LINE=solid; CMI=1; LABEL="AIFM";;
+        "fswap")                pattern="04-11"; rmem=fastswap; backend=rdma; cores=${CORES}; evb=; evp=; evprio=; desc="rdma"; LS=solid; CMI=1; LABEL="Fastswap";;
+        "eden-sc")              pattern="04-11"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=yes; desc=${EDENDESC}; LS=solid; CMI=0; LABEL="Eden(Prio)";;
+        "eden+bh-sc")           pattern="04-\(09\|10\|11\)"; rmem=eden-bh; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=yes; desc=${EDENDESC}; LS=dashed; CMI=1; LABEL="Eden(BH,Prio)";;
+        "eden")                 pattern="04-\(09\|10\|11\)"; rmem=eden; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc=${EDENDESC}; LS=solid; CMI=0; LABEL="Eden(Prio,SC)";;
+        "eden+bh")              pattern="04-\(09\|10\|11\)"; rmem=eden-bh; backend=rdma; cores=${CORES}; evb=32; evp=SC; evprio=yes; desc=${EDENDESC}; LS=dashed; CMI=1; LABEL="Eden(BH,SC,Prio)";;
+        "eden+bh-sc-prio")      pattern="04-\(09\|10\|11\)"; rmem=eden-bh; backend=rdma; cores=${CORES}; evb=32; evp=NONE; evprio=no; desc=${EDENDESC}; LS=dashdot; CMI=1; LABEL="Eden(Basic)";;
         *)                      echo "Unknown config"; exit;;
         esac
 
         # filter results
         label=$runcfg
-        datafile=$plotdir/data_${LABEL}
+        datafile=$plotdir/data_${LABEL}_zs${ZIPFS}
         descopt=
         evbopt=
         rmemopt=
@@ -134,35 +155,52 @@ if [ "$PLOTID" == "2" ]; then
         if [[ $evprio ]];  then evpropt="-evpr=$evprio";    fi
 
         if [[ $FORCE ]] || [ ! -f "$datafile" ]; then
-            echo "LMem%,Xput,XputErr,Faults,FaultsErr,NetReads,NetReadsErr,HitR,Count,System,Backend,EvP,EvB,CPU,Zipfs" > $datafile
-            for memp in 100 91 83 75 66 58 50 41 33 22 16 8 4; do
-                tmpfile=${TMP_FILE_PFX}data
-                rm -f ${tmpfile}
-                echo bash ${SCRIPT_DIR}/show.sh -cs="$pattern" -be=$backend -c=$cores -of=$tmpfile  \
-                    -be=${bkend} ${descopt} ${evbopt} ${rmemopt} ${evpopt} ${rdopt} ${evpropt} -lmp=${memp}
-                bash ${SCRIPT_DIR}/show.sh -cs="$pattern" -be=$backend -c=$cores -of=$tmpfile  \
-                    -be=${bkend} ${descopt} ${evbopt} ${rmemopt} ${evpopt} ${rdopt} ${evpropt} -lmp=${memp}
-                cat $tmpfile
-                xmean=$(csv_column_mean $tmpfile "Xput")
-                xstd=$(csv_column_stdev $tmpfile "Xput")
-                xnum=$(csv_column_count $tmpfile "Xput")
-                fmean=$(csv_column_mean $tmpfile "Faults")
-                fstd=$(csv_column_stdev $tmpfile "Faults")
-                nrmean=$(csv_column_mean $tmpfile "NetReads")
-                nrstd=$(csv_column_stdev $tmpfile "NetReads")
-                hitrmean=$(csv_column_mean $tmpfile "HitR")
-                echo ${memp},${xmean},${xstd},${fmean},${fstd},${nrmean},${nrerr},${hitrmean},${xnum},${rmem},${bkend},${evp},${evb},${cores},${zipfs} >> ${datafile}
-            done
+            if [[ "$rmem" == "aifm" ]]; then
+                if [[ ! -f "$datafile" ]]; then
+                    echo "ERROR! Expecting AIFM data in $datafile"
+                    exit 1
+                fi
+            else
+                echo "LMem%,Xput,XputErr,Faults,FaultsErr,NetReads,NetReadsErr,KFaults,KFaultsErr,HitR,Count,System,Backend,EvP,EvB,CPU,Zipfs" > $datafile
+                for memp in 100 91 83 75 66 58 50 41 33 22 16 8 4; do
+                    tmpfile=${TMP_FILE_PFX}data
+                    rm -f ${tmpfile}
+                    echo bash ${SCRIPT_DIR}/show.sh -cs="$pattern" -be=$backend -c=$cores -of=$tmpfile  \
+                        ${descopt} ${evbopt} ${rmemopt} ${evpopt} ${rdopt} ${evpropt} -zs=${ZIPFS} -lmp=${memp}
+                    bash ${SCRIPT_DIR}/show.sh -cs="$pattern" -be=$backend -c=$cores -of=$tmpfile  \
+                        ${descopt} ${evbopt} ${rmemopt} ${evpopt} ${rdopt} ${evpropt} -zs=${ZIPFS} -lmp=${memp}
+                    cat $tmpfile
+                    xmean=$(csv_column_mean $tmpfile "Xput")
+                    xstd=$(csv_column_stdev $tmpfile "Xput")
+                    xnum=$(csv_column_count $tmpfile "Xput")
+                    fmean=$(csv_column_mean $tmpfile "Faults")
+                    fstd=$(csv_column_stdev $tmpfile "Faults")
+                    nrmean=$(csv_column_mean $tmpfile "NetReads")
+                    nrstd=$(csv_column_stdev $tmpfile "NetReads")
+                    kfmean=$(csv_column_mean $tmpfile "KFaults")
+                    kfstd=$(csv_column_stdev $tmpfile "KFaultsErr")
+                    hitrmean=$(csv_column_mean $tmpfile "HitR")
+                    echo ${memp},${xmean},${xstd},${fmean},${fstd},${nrmean},${nrerr},${kfmean},${kfstd},${hitrmean},${xnum},${rmem},${bkend},${evp},${evb},${cores},${zipfs} >> ${datafile}
+                done
 
-            # compute and add normalized throughput column
-            maxput=$(csv_column_max "$datafile" "Xput")
-            echo $maxput
-            xputnorm=$(csv_column "$datafile" "Xput" | awk '{ if($1) print $1/'$maxput'; else print ""; }')
-            xputerrnorm=$(csv_column "$datafile" "XputErr" | awk '{ if($1) print $1/'$maxput'; else print ""; }')
-            echo -e "XputNorm\n${xputnorm}" > ${TMP_FILE_PFX}normxput
-            echo -e "XputErrNorm\n${xputerrnorm}" > ${TMP_FILE_PFX}normxputerr
-            paste -d, $datafile ${TMP_FILE_PFX}normxput ${TMP_FILE_PFX}normxputerr > ${TMP_FILE_PFX}normdata
-            mv ${TMP_FILE_PFX}normdata $datafile
+                # compute and add normalized throughput column
+                case "$rmem-$ZIPFS" in
+                "eden-0.85")           maxput=526952;;
+                "eden-bh-0.85")        maxput=526952;;
+                "fastswap-0.85")       maxput=518883;;
+                "eden-0.5")            maxput=514768;;
+                "eden-bh-0.5")         maxput=514768;;
+                "fastswap-0.5")        maxput=490191;;
+                *)          echo "Unknown rmem"; exit;;
+                esac
+                echo $maxput
+                xputnorm=$(csv_column "$datafile" "Xput" | awk '{ if($1) print $1/'$maxput'; else print ""; }')
+                xputerrnorm=$(csv_column "$datafile" "XputErr" | awk '{ if($1) print $1/'$maxput'; else print ""; }')
+                echo -e "XputNorm\n${xputnorm}" > ${TMP_FILE_PFX}normxput
+                echo -e "XputErrNorm\n${xputerrnorm}" > ${TMP_FILE_PFX}normxputerr
+                paste -d, $datafile ${TMP_FILE_PFX}normxput ${TMP_FILE_PFX}normxputerr > ${TMP_FILE_PFX}normdata
+                mv ${TMP_FILE_PFX}normdata $datafile
+            fi
         fi
 
         label=${LABEL:-$runcfg}
@@ -186,25 +224,36 @@ if [ "$PLOTID" == "2" ]; then
         YLABEL="Normalized Throughput"
         YMUL=
     fi
-    plotname=${plotdir}/synthetic_prio_xput.${PLOTEXT}
+    plotname=${plotdir}/synthetic_prio_xput_${cfg}_zs${ZIPFS}.${PLOTEXT}
     if [[ $FORCE_PLOTS ]] || [ ! -f "$plotname" ]; then
         python3 ${ROOTDIR}/scripts/plot.py ${plots}             \
             -yce ${XPUTCOL} ${XPUTERR} -yl "${YLABEL}" ${YMUL} ${YLIMS}     \
             -xc "LMem%" -xl "Local Mem (%)"                     \
-            --size 4.5 3 -fs 10 -of $PLOTEXT -o $plotname
+            --size 6 6 -fs 11 -of $PLOTEXT -o $plotname
     fi
     files="$files $plotname"
 
     #plot faults
-    YLIMS="--ymin 0 --ymax 850"
-    plotname=${plotdir}/synthetic_netreads.${PLOTEXT}
+    YLIMS="--ymin 0 --ymax 1300"
+    plotname=${plotdir}/synthetic_netreads_${cfg}_zs${ZIPFS}.${PLOTEXT}
     if [[ $FORCE_PLOTS ]] || [ ! -f "$plotname" ]; then
         python3 ${ROOTDIR}/scripts/plot.py ${plots}                     \
             -yce "NetReads" "NetReadsErr" -yl "Remote Page Fetches(KOPS)" --ymul 1e-3 ${YLIMS}   \
             -xc "LMem%" -xl "Local Mem (%)" ${XLIMS}                    \
-            --size 4.5 3 -fs 10 -of $PLOTEXT -o $plotname
+            --size 6 6 -fs 11 -of $PLOTEXT -o $plotname
     fi
     files="$files $plotname"
+
+    # #plot unhinted faults
+    # YLIMS="--ymin 0 --ymax 200"
+    # plotname=${plotdir}/synthetic_kfaults.${PLOTEXT}
+    # if [[ $FORCE_PLOTS ]] || [ ! -f "$plotname" ]; then
+    #     python3 ${ROOTDIR}/scripts/plot.py ${plots}                     \
+    #         -yce "KFaults" "KFaultsErr" -yl "Unhinted Faults (KOPS)" --ymul 1e-3 ${YLIMS}   \
+    #         -xc "LMem%" -xl "Local Mem (%)" ${XLIMS}                    \
+    #         --size 4.5 3 -fs 10 -of $PLOTEXT -o $plotname
+    # fi
+    # files="$files $plotname"
 
     # ## Hit ratio
     # YLIMS="--ymin 0 --ymax 100"
@@ -218,8 +267,8 @@ if [ "$PLOTID" == "2" ]; then
     # files="$files $plotname"
 
     # Combine
-    plotname=${plotdir}/${cfg}.$PLOTEXT
-    montage -tile 3x0 -geometry +5+5 -border 5 $files ${plotname}
+    plotname=${plotdir}/${cfg}_zs${ZIPFS}.$PLOTEXT
+    montage -tile 2x0 -geometry +5+5 -border 5 $files ${plotname}
     display ${plotname} &
 fi
